Move DummyIMBackend to fixtures
